Inline snippets in the test
