--- detekt-rules-empty/src/test/resources/EmptyDefaultConstructorNegative.kt
@@ -1,14 +1,0 @@
-@file:Suppress("unused", "UNUSED_PARAMETER")
-
-package cases
-
-class PrimaryConstructorWithParameter constructor(x: Int)
-
-class PrimaryConstructorWithAnnotation @SafeVarargs constructor()
-
-class PrivatePrimaryConstructor private constructor()
-
-class EmptyConstructorIsCalled() {
-
-    constructor(i: Int) : this()
-}
