--- detekt-rules-empty/src/test/resources/EmptyDefaultConstructorPositive.kt
@@ -1,7 +1,0 @@
-@file:Suppress("unused")
-
-package cases
-
-class EmptyPrimaryConstructor constructor()
-
-class EmptyPublicPrimaryConstructor public constructor()
